Split dig.go into multiple files. Only leave interface
